Hide game list in debug mode, Hide menu bar when not in debug mode
The menu bar will be hidden if we're not in debug mode. The game list will be hidden when not in debug mode. This allows more screen space for debugging. Additionally, the game list soon won't be used to launch a game, so having it accessible in debug mode won't be necessary